Jk/bug cumulus 1 fix local api (#3374)
* Additional fixes in support of existing PR

* Add lint override for shebang

* Remove un-needed cli option

* Fixup
