+++ packages/api/bin/cli.js
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+/* eslint-disable node/shebang */
 
 'use strict';
